Reformat `dataPotions.js` to make each potion entry clearer
There is now a blank line between each potion entry, which makes it easier to distinguish between separate entries and entries where the player gets to choose between multiple potions.

This made it easier to identify and fix a syntax error with the entry that now starts on line 131. Some of the closing brackets on its `params` were missing.